refactor(blorktools): loading-splash.js
