+++ packages/blorktools/src/asset_debugger/scene/asset_debugger.js
@@ -29,6 +29,7 @@ import { initWorldPanel } from '../panels/world-panel/world-panel.js';
 import { initState } from './state.js';
 import { initUiManager } from '../util/ui-manager.js';
 import { setupDropzones } from '../landing-page/file-handler.js';
+import { hideLoadingSplash, showLoadingSplash, updateLoadingProgress } from '../loading-splash/loading-splash.js';
 
 // Debug flags
 const DEBUG_LIGHTING = false;
@@ -1097,63 +1098,7 @@ document.addEventListener('DOMContentLoaded', () => {
     }
 });
 
-// LOADING FUNCTIONS
 
-/**
- * Shows the loading splash screen
- */
-function showLoadingSplash() {
-    // First check if the splash already exists
-    let loadingSplash = document.getElementById('loading-splash');
-    
-    if (!loadingSplash) {
-        // Create and add the loading splash screen from our HTML template
-        fetch('../loading-splash/loading-splash.html')
-            .then(response => response.text())
-            .then(html => {
-                document.body.insertAdjacentHTML('beforeend', html);
-                // Make sure it's visible
-                loadingSplash = document.getElementById('loading-splash');
-                if (loadingSplash) {
-                    loadingSplash.style.display = 'flex';
-                }
-            })
-            .catch(error => {
-                console.error('Error loading splash screen:', error);
-            });
-    } else {
-        // If it exists, make sure it's visible
-        loadingSplash.style.display = 'flex';
-        loadingSplash.classList.remove('fade-out');
-    }
-}
-
-/**
- * Updates the loading progress text on the splash screen
- * @param {string} text - The progress message to display
- */
-function updateLoadingProgress(text) {
-    const progressText = document.getElementById('loading-progress-text');
-    if (progressText) {
-        progressText.textContent = text;
-    }
-}
-
-/**
- * Hides the loading splash screen with a fade-out animation
- */
-function hideLoadingSplash() {
-    const loadingSplash = document.getElementById('loading-splash');
-    if (loadingSplash) {
-        // Add fade-out class for smooth transition
-        loadingSplash.classList.add('fade-out');
-        
-        // Remove the element after the animation completes
-        setTimeout(() => {
-            loadingSplash.remove();
-        }, 600); // Match the transition duration in CSS
-    }
-}
 
 /**
  * Checks if all resources have loaded and hides the splash screen when done
